+++ 017_r_use_nlme.R
@@ -19,7 +19,6 @@
 #'     code_folding: show
 #' ---
 
-
 # Prepare environment ---- 
 
 # _1.) Collect garbage ----
@@ -29,12 +28,14 @@ gc()
 
 # _2.) Packages ----
 
-library("here")
-library("dplyr")
-library("ggplot2")
+library("here")    # handle path names
+library("dplyr")   # handle data more easily
 
-library("nlme")
-library("performance") # Model inspection
+library("lattice") # create trellis graphs
+library("ggpubr")  # save trellis graphs
+library("ggplot2") # save trellis graphs
+
+library("nlme")    # model non-linear mixed-effect dependencies
 
 # Setup data and model ----
 
@@ -49,30 +50,41 @@ mice_f1_slct <- left_join(mice_f1_slct, {mice_f0_slct %>% dplyr::select(AnimalId
 
 # _3.) Check data ----
 
-mice_f1_slct %>% dplyr::select(MeasurementDay, BodyWeightG, AnimalId, AnimalSex) %>% arrange(AnimalSex, AnimalId, MeasurementDay) %>% print(n = Inf)
+plot_data_check <- xyplot(BodyWeightG ~ MeasurementDay | AnimalId, groups = AnimalSex, data = mice_f1_slct, xlab = "day [d]", ylab = "body weight [g]", auto.key = list(title = "sex"))
+plot_data_check
 
-ggplot(data = mice_f1_slct, aes(x = "MeasurementDay", y="BodyWeightG",  group = "AnimalId")) +
-  geom_line( mapping = aes( x = MeasurementDay, y = BodyWeightG, group = AnimalId, color = AnimalId)) + 
-  facet_grid(. ~ AnimalSex) +
-  theme_bw()
+ggsave("017_r_use_nlme__data_check.pdf", plot = ggarrange(plot_data_check), path = here("../manuscript/display_items"),
+  scale = 1, width = 9, height = 5, units = c("in"), dpi = 300, limitsize = TRUE)
 
 # _4.) Define possibly applicable model functions ----
 
-# __a) Exponential approach as in {015_r_use_saemix.R}  ----
+# __a) Exponential approach as in "015_r_use_saemix.R" ----
 
-decay.formula <-  as.formula(y ~ a * (1 - b * exp( -k * x)))
+approach.formula <- as.formula(y ~ a * (1 - b * exp( -k * x)))
     
 # __b) Testing function to get starting values ----
 
 # from estimated previous values - equivalent model (RQ1)
-a = 25.3005
-b = 1.2644
-k = 0.0392
-curve(a * (1 - b * exp( -k * x)), from = 35, to = 100, xlab="MeasurementDay", ylab="BodyWeightG", col = "darkgray", xlim =c(30, 100), ylim=c(15, 25))
+
+a = 25.3005; b = 1.2644; k = 0.0392
+curve(a * (1 - b * exp( -k * x)), from = 35, to = 100, xlab = "day [d]", ylab = "body weight [g]", col = "darkgray", xlim =c(30, 100), ylim=c(15, 25))
 
 # RQ1: What are the effects of diet regardless of sex ? ----
 
-# _1.) Get a suitable null model which is comparable to the subsequent one ----
+# _1.) Get a suitable null model ----
+
+# __a) Intended model:
+
+plot_rq1_null_model <- xyplot(BodyWeightG ~ MeasurementDay | AnimalSex, groups = AnimalId , data = mice_f1_slct, xlab = "day [d]", ylab = "body weight [g]",
+                              panel = function(x, y) {
+                                panel.xyplot(x, y)
+                                panel.loess(x, y)
+                                })
+
+ggsave("017_r_use_nlme__rq1_null_model.pdf", plot = ggarrange(plot_rq1_null_model), path = here("../manuscript/display_items"),
+       scale = 1, width = 5, height = 5, units = c("in"), dpi = 300, limitsize = TRUE)
+
+# __a) Build model:
 
 exp.appr.fit.diet.nosex.null <- nlme(BodyWeightG ~ a * (1 - b * exp( -k * MeasurementDay)),
                                      data = mice_f1_slct,
@@ -80,77 +92,124 @@ exp.appr.fit.diet.nosex.null <- nlme(BodyWeightG ~ a * (1 - b * exp( -k * Measur
                                      random = a  ~  AnimalSex | AnimalId,
                                      na.action = na.exclude,
                                      start = c(25.30,  1.31,  0.04),
-                                     control = nlmeControl(msMaxIter = 1000, msVerbose = TRUE))
+                                     control = nlmeControl(msMaxIter = 50, msVerbose = FALSE))
 
 # _2.) Check null model ----
 
 summary(exp.appr.fit.diet.nosex.null)
+
 # AIC      BIC    logLik
 # 932.4092 958.3356 -459.2046
 
-# _3.) Plot null model residuals ----
+# Fixed effects:  a + b + k ~ 1 
+# Value  Std.Error  DF  t-value p-value
+# a.(Intercept) 23.112100 0.25354413 248 91.15613       0
+# b              1.311560 0.06272352 248 20.91018       0
+# k              0.040515 0.00176842 248 22.91022       0
+
+# _3.) Plot null model fits  ----
+
+plot_rq1_null_model_fit <- xyplot(BodyWeightG ~ MeasurementDay | AnimalId, data = mice_f1_slct, fit = exp.appr.fit.diet.nosex.null,
+       strip = TRUE, aspect = "xy", grid = TRUE,
+       panel = function(x, y, ..., fit, subscripts) {
+         panel.xyplot(x, y, ...)
+         ypred <- fitted(fit)[subscripts]
+         panel.lines(x, ypred, col = "black")
+       },
+       xlab = "day [d]", ylab = "body weight [g]")
+
+ggsave("017_r_use_nlme__plot_rq1_null_model_fit.pdf", plot = ggarrange(plot_rq1_null_model_fit), path = here("../manuscript/display_items"),
+       scale = 1, width = 12, height = 5, units = c("in"), dpi = 300, limitsize = TRUE)
+
+# _4.) Plot null model residuals ----
 
 plot(exp.appr.fit.diet.nosex.null) # residuals seem ok
 
-# _4.) Get diet model as in RQ4 in `015_r_use_saemix.R` but without litter size ----
+# ** continue here **
 
-# Diet interactions were not significant
+# _5.) Get matching diet model as in RQ4 in `015_r_use_saemix.R` but without litter size ----
+
+# (Diet interactions were not significant)
 
 exp.appr.fit.diet.nosex <- nlme(BodyWeightG ~ a * (1 - b * exp( -k * MeasurementDay)),
                                 data = mice_f1_slct,
                                 fixed  = a + b + k ~  FatherDiet + MotherDiet,
-                                random = a  ~ AnimalSex | AnimalId,
+                                random = a  ~  AnimalSex | AnimalId,
                                 na.action = na.exclude,
                                 start = c(25.30,  1.31,  0.04,
                                           0.17,  0.09,  0.04,
                                           0.01,  0.08,  0.02),
-                                control = nlmeControl(msMaxIter = 1000, msVerbose = TRUE))
+                                control = nlmeControl(msMaxIter = 50, msVerbose = FALSE))
 
 # _5.) Check diet model ----
 
 summary(exp.appr.fit.diet.nosex)
-# AIC      BIC    logLik
-# 924.2106 972.3597 -449.1053
 
-# _6.) Plot diet model residuals ----
+# AIC      BIC    logLik
+# 924.2106 972.3597 -449.1053 <- better then null model - see LRT below
+
+# Fixed effects:  a + b + k ~ FatherDiet + MotherDiet 
+# Value Std.Error  DF  t-value p-value
+# a.(Intercept)   24.262459 0.4877282 242 49.74586  0.0000 *
+# a.FatherDietHFD -1.037515 0.4913831 242 -2.11142  0.0358
+# a.MotherDietHFD -0.943602 0.4729634 242 -1.99508  0.0472 * only a.MotherDietHFD significant
+# b.(Intercept)    1.329206 0.1167432 242 11.38572  0.0000
+# b.FatherDietHFD  0.001529 0.1247160 242  0.01226  0.9902
+# b.MotherDietHFD -0.030819 0.1234833 242 -0.24958  0.8031
+# k.(Intercept)    0.038565 0.0032705 242 11.79192  0.0000
+# k.FatherDietHFD  0.001229 0.0035098 242  0.35004  0.7266
+# k.MotherDietHFD  0.002341 0.0034729 242  0.67419  0.5008
+
+# _6.) Plot diet model fits  ----
+
+plot_rq1_diet_model_fit <- xyplot(BodyWeightG ~ MeasurementDay | AnimalId, data = mice_f1_slct, fit = exp.appr.fit.diet.nosex,
+                                  strip = TRUE, aspect = "xy", grid = TRUE,
+                                  panel = function(x, y, ..., fit, subscripts) {
+                                    panel.xyplot(x, y, ...)
+                                    ypred <- fitted(fit)[subscripts]
+                                    panel.lines(x, ypred, col = "black")
+                                  },
+                                  xlab = "day [d]", ylab = "body weight [g]")
+
+ggsave("017_r_use_nlme__plot_rq1_diet_model_fit.pdf", plot = ggarrange(plot_rq1_diet_model_fit), path = here("../manuscript/display_items"),
+       scale = 1, width = 12, height = 5, units = c("in"), dpi = 300, limitsize = TRUE)
+
+# _7.) Plot diet model residuals ----
 
 plot(exp.appr.fit.diet.nosex) # residuals seem ok
 
-# _7.) Compare models ----
+# _8.) Compare models ----
 
-anova(exp.appr.fit.diet.nosex.null, exp.appr.fit.diet.nosex) # adding diet imporves model
+anova(exp.appr.fit.diet.nosex.null, exp.appr.fit.diet.nosex) # adding diet improves model
 
-# _8.) Plot model predictions ----
+# ** Continue here **
 
-# __a) Curve plots ----
+# _9.) Plot diet model predictions ----
 
-# null model curve - all mice ragrdless of sex
+# null model curve - all mice regardless of sex
 a = fixed.effects(exp.appr.fit.diet.nosex.null)[1]
 b = fixed.effects(exp.appr.fit.diet.nosex.null)[2]
 k = fixed.effects(exp.appr.fit.diet.nosex.null)[3]
-curve(a * (1 - b * exp( -k * x)), from = 35, to = 100, xlab="MeasurementDay", ylab="BodyWeightG", col = "darkgray", xlim =c(30, 100), ylim=c(15, 25))
+curve(a * (1 - b * exp( -k * x)), from = 35, to = 100, xlab = "day [d]", ylab = "body weight [g]", col = "darkgray", xlim =c(30, 100), ylim=c(15, 25))
 
-
-# chow diet curve  - all mice ragrdless of sex
+# chow diet curve  - all mice regardless of sex
 a = fixed.effects(exp.appr.fit.diet.nosex)[1]
 b = fixed.effects(exp.appr.fit.diet.nosex)[4]
 k = fixed.effects(exp.appr.fit.diet.nosex)[7]
-curve(a * (1 - b * exp( -k * x)), from = 35, to = 100, xlab="MeasurementDay", ylab="BodyWeightG", col = "black", xlim =c(30, 100), ylim=c(15, 25), add = TRUE)
+curve(a * (1 - b * exp( -k * x)), from = 35, to = 100, xlab = "day [d]", ylab = "body weight [g]", col = "black", xlim =c(30, 100), ylim=c(15, 25), add = TRUE)
 
-# father hfd diet curve  - all mice ragrdless of sex
+# father hfd diet curve  - all mice regardless of sex
 
 af = a + fixed.effects(exp.appr.fit.diet.nosex)[2] # * 
 bf = b + fixed.effects(exp.appr.fit.diet.nosex)[5]
 kf = k + fixed.effects(exp.appr.fit.diet.nosex)[8]
-curve(af * (1 - bf * exp( -kf * x)), from = 35, to = 100, xlab="MeasurementDay", ylab="BodyWeightG", col = "darkorange", xlim =c(30, 100), ylim=c(15, 25), add = TRUE)
+curve(af * (1 - bf * exp( -kf * x)), from = 35, to = 100, xlab = "day [d]", ylab = "body weight [g]", col = "darkorange", xlim =c(30, 100), ylim=c(15, 25), add = TRUE)
 
-# mother hfd diet curve  - all mice ragrdless of sex
+# mother hfd diet curve  - all mice regardless of sex
 af = a + fixed.effects(exp.appr.fit.diet.nosex)[3] # * 
 bf = b + fixed.effects(exp.appr.fit.diet.nosex)[6]
 kf = k + fixed.effects(exp.appr.fit.diet.nosex)[9]
-curve(af * (1 - bf * exp( -kf * x)), from = 35, to = 100, xlab="MeasurementDay", ylab="BodyWeightG", col = "red", xlim =c(30, 100), ylim=c(15, 25), add = TRUE)
-
-# __b) Prediction plots ----
+curve(af * (1 - bf * exp( -kf * x)), from = 35, to = 100, xlab = "day [d]", ylab = "body weight [g]", col = "red", xlim =c(30, 100), ylim=c(15, 25), add = TRUE)
 
 # RQ2: Get a null model to investigate the sex specific effcet more ----
 
@@ -303,7 +362,6 @@ a = fixed.effects(exp.appr.fit.diet)[1] + fixed.effects(exp.appr.fit.diet)[1+1]
 b = fixed.effects(exp.appr.fit.diet)[5] + fixed.effects(exp.appr.fit.diet)[5+1]
 k = fixed.effects(exp.appr.fit.diet)[9] + fixed.effects(exp.appr.fit.diet)[9+1]
 curve(a * (1 - b * exp( -k * x)), from = 35, to = 100, xlab="MeasurementDay", ylab="BodyWeightG", col = "darkgrey", xlim =c(30, 100), ylim=c(15, 25), add = TRUE)
-
 
 # fathers hfd diet males
 a = fixed.effects(exp.appr.fit.diet)[1] + fixed.effects(exp.appr.fit.diet)[1+1] + fixed.effects(exp.appr.fit.diet)[1+2]